backend/requirements.txt
Django==3.0.5
django-cors-headers==3.2.0
django-ordered-model==3.3.0
djangorestframework==3.10.3
gunicorn==20.0.4
Markdown==3.1.1
Pillow==6.2.1
psycopg2-binary==2.8.6
pytz==2019.3
PyYAML==5.3.1
python-dotenv==0.12.0

pep8==1.7.1
autopep8==1.4.4
pycodestyle==2.5.0
requests==2.23.0
